refactor(BookContent): 리스트 상세 영역 컴포넌트로 분리
