+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './shared';
+export { default as BookContext } from './BookContext';
 export { default as BookList } from './BookList';
 export { default as DetailSearch } from './DetailSearch';
 export { default as LikedList } from './LikedList';
